made functions static, modified service

+++ src/main/java/ee/taltech/calculator/dto/CalculationResult.java
@@ -11,5 +11,5 @@ public class CalculationResult {
     private Integer minOdd;
     private List<Integer> even;
     private List<Integer> squared;
-    private Float averageOfPositives;
+    private Double averageOfPositives;
 }
